New feature : Add Product, Add Get All Specification, Add Upload Image Module

+++ src/main/java/com/swp/whmsystem/model/Chip.java
@@ -2,16 +2,15 @@ package com.swp.whmsystem.model;
 
 public class Chip {
 
-    private Long id;
-
+    private int id;
     private String name;
     private boolean active;
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
